+++ srcs/expansion/expand_cmds.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expand_cmds.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 14:25:22 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 14:25:24 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
 static int	process_expand_cmds(t_exp_tkn **head, char **argv)
@@ -7,7 +19,6 @@ static int	process_expand_cmds(t_exp_tkn **head, char **argv)
 	i = 0;
 	while (argv[i])
 	{
-		// 変数展開とクォート除去
 		if (expand_unsplit(head, argv[i], 1) != 0)
 		{
 			free_exp_tokens(*head);
@@ -18,7 +29,6 @@ static int	process_expand_cmds(t_exp_tkn **head, char **argv)
 	return (0);
 }
 
-// コマンドのargvの展開
 int	expand_cmds(t_node *node)
 {
 	t_exp_tkn	*head;
@@ -26,7 +36,6 @@ int	expand_cmds(t_node *node)
 	head = NULL;
 	if (process_expand_cmds(&head, node->argv) != 0)
 		return (-1);
-	// 変数展開の結果を単語分割
 	if (split_exp_tokens(&head) != 0)
 	{
 		free_exp_tokens(head);
