+++ srcs/expansion/expand_redirs.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expand_redirs.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 14:27:39 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 14:27:45 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
 static int	split_expanded_redir_env(t_exp_tkn **head, char *str)
@@ -5,13 +17,11 @@ static int	split_expanded_redir_env(t_exp_tkn **head, char *str)
 	int			count;
 	t_exp_tkn	*cur;
 
-	// 変数展開の結果を単語分割
 	if (split_exp_tokens(head) != 0)
 	{
 		free_exp_tokens(*head);
 		return (1);
 	}
-	// 変数展開→単語分割後の文字列の数を数える
 	count = 0;
 	cur = *head;
 	while (cur)
@@ -19,7 +29,6 @@ static int	split_expanded_redir_env(t_exp_tkn **head, char *str)
 		count++;
 		cur = cur->next;
 	}
-	// 文字列が複数ならambiguous redirectエラー
 	if (count > 1)
 	{
 		ft_dprintf(STDERR_FILENO, "minishell: %s: ambiguous redirect\n", str);
@@ -54,7 +63,6 @@ static int	process_expand_redirs(t_exp_tkn **new, t_redir *redir, int env_flag)
 	return (0);
 }
 
-// リダイレクトのファイル名またはデリミタの展開
 int	expand_redirs(t_node *node)
 {
 	int			i;
@@ -66,7 +74,6 @@ int	expand_redirs(t_node *node)
 	{
 		new = NULL;
 		env_flag = 0;
-		// heredocでなければ$展開を行う
 		if (node->redirs[i]->kind != REDIR_HEREDOC)
 			env_flag = 1;
 		if (process_expand_redirs(&new, node->redirs[i], env_flag) != 0)
