+++ srcs/init/init.c
@@ -1,39 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 14:28:11 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 14:32:46 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-// SIGINTとSIGQUITに対してシグナルハンドラなどの設定
 void	init_signals(void)
 {
 	set_main_sigint();
 	set_main_sigquit();
-}
-
-char	**get_envp_copy(char **envp)
-{
-	char	**cp;
-	int		env_count;
-	int		i;
-
-	env_count = 0;
-	while (envp[env_count])
-		env_count++;
-	cp = (char **)ft_malloc(sizeof(char *) * (env_count + 1));
-	if (!cp)
-		return (NULL);
-	i = 0;
-	while (i < env_count)
-	{
-		cp[i] = ft_strdup(envp[i]);
-		if (!cp[i])
-		{
-			while (--i >= 0)
-				free(cp[i]);
-			free(cp);
-			return (NULL);
-		}
-		i++;
-	}
-	cp[i] = NULL;
-	return (cp);
 }
 
 int	init_env(char **envp)
@@ -79,77 +61,6 @@ int	init_pwd(void)
 	return (0);
 }
 
-int	pure_atoi(char *num)
-{
-	int	rtn;
-	int	i;
-	int	sign;
-
-	rtn = 0;
-	i = 0;
-	sign = 1;
-	if (ft_isspace(num[i]))
-		++i;
-	if (num[i] == '-')
-	{
-		sign = -1;
-		++i;
-	}
-	else if (num[i] == '+')
-		++i;
-	while (ft_isdigit(num[i]))
-	{
-		rtn *= 10;
-		rtn += num[i] - '0';
-		++i;
-	}
-	return (rtn * sign);
-}
-
-int	is_valid_shlvl(char *num)
-{
-	int	i;
-
-	i = 0;
-	while (ft_isspace(num[i]))
-		++i;
-	if (num[i] == '+' || num[i] == '-')
-		++i;
-	while (ft_isdigit(num[i]))
-		++i;
-	while (ft_isspace(num[i]))
-		++i;
-	if (num[i] == '\0')
-		return (0);
-	return (1);
-}
-
-int	increment_shlvl(int shlvl)
-{
-	char	*str_shlvl;
-	char	*env_shlvl;
-
-	str_shlvl = ft_itoa(shlvl + 1);
-	env_shlvl = ft_strjoin("SHLVL=", str_shlvl);
-	if (!str_shlvl || !env_shlvl)
-	{
-		free(str_shlvl);
-		free(env_shlvl);
-		ft_env(ENV_DEL_ALL, NULL);
-		return (1);
-	}
-	if (ft_env(ENV_SET, env_shlvl) == NULL)
-	{
-		free(str_shlvl);
-		free(env_shlvl);
-		ft_env(ENV_DEL_ALL, NULL);
-		return (1);
-	}
-	free(str_shlvl);
-	free(env_shlvl);
-	return (0);
-}
-
 int	init_shlvl(void)
 {
 	char	*shlvl;
@@ -163,21 +74,7 @@ int	init_shlvl(void)
 		return (0);
 	}
 	lvl = pure_atoi(shlvl);
-	if (lvl > 998)
-	{
-		ft_dprintf(STDERR_FILENO, "minishell: warning: shell level (%d) "
-			"too high, resetting to 1\n", lvl + 1);
-		if (ft_env(ENV_SET, "SHLVL=1") == NULL)
-			return (1);
-	}
-	else if (lvl < 0)
-	{
-		if (ft_env(ENV_SET, "SHLVL=0") == NULL)
-			return (1);
-	}
-	else if (increment_shlvl(lvl))
-		return (1);
-	return (0);
+	return (handle_shlvl_range(lvl));
 }
 
 int	init(char **envp)
