+++ srcs/interpret.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   interpret.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 14:14:25 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 14:14:27 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
 int	handle_stage_ret(int ret)
@@ -20,7 +32,7 @@ void	interpret(t_shell *shell)
 {
 	int	ret;
 
-    ret = 0;
+	ret = 0;
 	ret = tokenize(shell->input, &shell->tokens);
 	if (handle_stage_ret(ret) != 0)
 		return ;
@@ -29,11 +41,9 @@ void	interpret(t_shell *shell)
 	ret = parse(shell->tokens, &shell->ast);
 	if (handle_stage_ret(ret) != 0)
 		return ;
-	// debug_parser(shell->ast);
 	ret = expand(shell->ast);
 	if (handle_stage_ret(ret) != 0)
 		return ;
-	// debug_expand(shell->ast);
 	ret = execute(shell->ast);
 	if (handle_stage_ret(ret) != 0)
 		return ;
