+++ srcs/parsing/peek.c
@@ -1,6 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   peek.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 15:08:01 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 15:08:03 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-// トークンがTK_WORDかどうか確認
 int	peek_word(t_token *token)
 {
 	if (token->type == TK_WORD)
@@ -8,13 +19,10 @@ int	peek_word(t_token *token)
 	return (0);
 }
 
-// トークンがリダイレクト演算子かどうか確認
 int	peek_redir_op(t_token *token)
 {
-	// トークンが予約語でなければエラー
 	if (token->type != TK_RESERVED)
 		return (0);
-	// トークンがリダイレクト記号かどうか
 	if (ft_strncmp(token->str, ">>", 3) == 0 || ft_strncmp(token->str, ">",
 			2) == 0 || ft_strncmp(token->str, "<<", 3) == 0
 		|| ft_strncmp(token->str, "<", 2) == 0)
