+++ srcs/redirection/create_temp.c
@@ -1,6 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   create_temp.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 15:08:48 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 15:08:49 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-// ヒアドキュメントの一時ファイルを生成する
 int	create_temp_file(int index, char **temp_file)
 {
 	char	*index_str;
