+++ srcs/redirection/redirect.c
@@ -1,6 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   redirect.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 15:09:17 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 15:09:51 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-// ヒアドキュメントの適用
 int	apply_heredoc_redir(t_redir *redir)
 {
 	int	heredoc_fd;
@@ -21,7 +32,6 @@ int	apply_heredoc_redir(t_redir *redir)
 	return (0);
 }
 
-// 入出力（ヒアドキュメント以外）のリダイレクト適用
 int	apply_io_redir(t_redir *redir, int flags, int std_fd)
 {
 	int	fd;
@@ -43,7 +53,6 @@ int	apply_io_redir(t_redir *redir, int flags, int std_fd)
 	return (0);
 }
 
-// 適用するリダイレクトによって分岐
 int	process_redir(t_redir *redir)
 {
 	if (redir->kind == REDIR_IN)
@@ -58,11 +67,10 @@ int	process_redir(t_redir *redir)
 		return (apply_heredoc_redir(redir));
 }
 
-// リダイレクトの適用メイン処理
 int	apply_redirs(int redir_count, t_redir **redirs)
 {
 	int	i;
-	int ret;
+	int	ret;
 
 	i = 0;
 	ret = 0;
