+++ srcs/redirection/unlink_temp.c
@@ -1,6 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   unlink_temp.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/15 15:08:28 by stakada           #+#    #+#             */
+/*   Updated: 2025/07/15 15:08:29 by stakada          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-// ヒアドキュメントの一時ファイルをすべて削除
 void	unlink_all_temp(t_node *first)
 {
 	t_node	*cur;
