+++ srcs/utils/shell.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   shell.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yumiyao <yumiyao@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 04:46:57 by yumiyao           #+#    #+#             */
-/*   Updated: 2025/05/29 05:02:15 by yumiyao          ###   ########.fr       */
+/*   Updated: 2025/07/15 15:06:20 by stakada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
